feat: better support for functions defined by macros
A macro can create many functions defined on the same line, which makes the
line metrics more confusing. For some function the line count can be 0, but
there's no way to distinguish that from the line counts.

This changes the count so that any function call to a line counts towards the
LH values, and we only keep the largest count for the DA values.